refactor(ppc,runner): modularize monolithic loader and execution engines
