+++ src/loader/ppc/mod.rs
@@ -82,7 +82,6 @@ use crate::quickdraw::fonts::{
 use crate::quickdraw::text::{
     get_font_metrics, get_glyph, get_glyph_italic, get_underline_thickness, QuickDrawTextStyle,
 };
-use crate::trap::dispatch::key_map_key_is_down;
 use crate::trap::extended80::Extended80;
 use crate::trap::types::{decode_mac_roman, encode_mac_roman_lossy, Rect};
 use crate::trap::{pict, TrapDispatcher};
@@ -91,10 +90,25 @@ use ppc::{
     PpcAlignmentPolicy, PpcCpu, PpcException, PpcFetchHistogram, PpcFetchObserver, PpcImportAction,
     PpcMemory, PpcMemoryWriteObserver, PpcNativeReturnGpr3, PpcRunResult, PpcSectionMemSpan,
 };
-use serde::{Deserialize, Serialize};
 use std::cell::Cell;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::sync::OnceLock;
+
+pub mod graphics;
+pub mod imports;
+pub mod qd3d;
+pub mod quicktime;
+pub mod sound;
+pub mod sprockets;
+pub mod vfs;
+
+pub use graphics::*;
+pub use imports::*;
+pub use qd3d::*;
+pub use quicktime::*;
+pub use sound::*;
+pub use sprockets::*;
+pub use vfs::*;
 
 pub const PPC_CODE_BASE: u32 = 0x0100_0000;
 pub const PPC_IMPORT_TVECTOR_BASE: u32 = 0x01e0_0000;
@@ -417,12 +431,7 @@ const PPC_DSP_BACK_VIS_RGN: u32 = 0x0501_0600;
 const PPC_DSP_BACK_CLIP_RGN_HANDLE: u32 = 0x0501_0700;
 const PPC_DSP_BACK_CLIP_RGN: u32 = 0x0501_0800;
 const PPC_DSP_BACK_SCREEN_BASE: u32 = 0x0502_0000;
-const PPC_MAIN_SCREEN_WIDTH: u32 = REFERENCE_MACHINE_PROFILE.screen_width as u32;
-const PPC_MAIN_SCREEN_HEIGHT: u32 = REFERENCE_MACHINE_PROFILE.screen_height as u32;
 const PPC_MAIN_PIXEL_DEPTH: u32 = REFERENCE_MACHINE_PROFILE.screen_depth as u32;
-const PPC_MAIN_SCREEN_STORAGE_DEPTH: u32 = 16;
-const PPC_DSP_SCREEN_WIDTH: u32 = 640;
-const PPC_DSP_SCREEN_HEIGHT: u32 = 480;
 pub const PPC_QD_TEXT_FONT_DEFAULT: i16 = 0;
 pub const PPC_QD_TEXT_MODE_SRC_OR: i16 = 1;
 pub const PPC_QD_TEXT_SIZE_SYSTEM: i16 = 0;
@@ -574,7 +583,6 @@ const PPC_Q3_GROUP_TYPE_IO_PROXY_DISPLAY: u32 = u32::from_be_bytes(*b"iopx");
 const PPC_Q3_GROUP_TYPE_ORDERED_DISPLAY: u32 = u32::from_be_bytes(*b"ordg");
 const PPC_Q3_GROUP_TYPE_LIGHT: u32 = u32::from_be_bytes(*b"lghg");
 const PPC_Q3_TYPE_ATTRIBUTE_SET: u32 = u32::from_be_bytes(*b"attr");
-const PPC_Q3_ILLUMINATION_TYPE_PHONG: u32 = u32::from_be_bytes(*b"phil");
 const PPC_Q3_ILLUMINATION_TYPE_LAMBERT: u32 = u32::from_be_bytes(*b"lmil");
 const PPC_Q3_ILLUMINATION_TYPE_NULL: u32 = u32::from_be_bytes(*b"nuil");
 const PPC_Q3_ATTRIBUTE_TYPE_AMBIENT_COEFFICIENT_3DMF: u32 = u32::from_be_bytes(*b"camb");
@@ -592,7 +600,6 @@ const PPC_Q3_PIXEL_TYPE_RGB16_565: u32 = 4;
 const PPC_Q3_PIXEL_TYPE_RGB24: u32 = 5;
 const PPC_Q3_ENDIAN_BIG: u32 = 0;
 const PPC_Q3_ENDIAN_LITTLE: u32 = 1;
-const PPC_Q3_ERROR_NONE: u32 = 0;
 const PPC_Q3_ERROR_INVALID_OBJECT_PARAMETER: u32 = 1;
 const PPC_Q3_VIEW_STATUS_DONE: u32 = 0;
 const PPC_Q3_VIEW_STATUS_ERROR: u32 = 2;
@@ -722,19 +729,14 @@ const PPC_Q3_ATTRIBUTE_TYPE_TRANSPARENCY_COLOR: u32 = 8;
 const PPC_Q3_ATTRIBUTE_TYPE_SURFACE_TANGENT: u32 = 9;
 const PPC_Q3_ATTRIBUTE_TYPE_HIGHLIGHT_STATE: u32 = 10;
 const PPC_Q3_ATTRIBUTE_TYPE_SURFACE_SHADER: u32 = 11;
-const PPC_Q3_SHADER_UV_BOUNDARY_WRAP: u32 = 0;
 const PPC_Q3_SHADER_UV_BOUNDARY_CLAMP: u32 = 1;
 const PPC_Q3_MATRIX3X3_SIZE: u32 = 36;
 const PPC_Q3_MATRIX4X4_SIZE: u32 = 64;
 const PPC_DSP_CONTEXT: u32 = 0x0500_0000;
 const PPC_DSP_CONTEXT_ATTRIBUTES_SIZE: u32 = 72;
 const PPC_DSP_DEPTH_MASK_8: u32 = 1 << 3;
-const PPC_DSP_DEPTH_MASK_16: u32 = 1 << 4;
-const PPC_DSP_CONTEXT_OPTION_QD3D_ACCEL: u32 = 1 << 0;
-const PPC_DSP_FREQUENCY_60HZ: u32 = 60 << 16;
 const PPC_DSP_DISPLAY_ID: u32 = 1;
 const PPC_DSP_BUFFER_KIND_NORMAL: u32 = 0;
-const PPC_DSP_ADVERTISED_PAGE_COUNT: u32 = 2;
 const PPC_DSP_MAX_REQUESTED_PAGE_COUNT: u32 = 2;
 const PPC_QA_ENGINE: u32 = 0x0500_0100;
 const PPC_QA_DRAW_CONTEXT: u32 = 0x0500_0110;
@@ -874,7 +876,6 @@ const PPC_FIXED_MAC_TIME: u32 = 3_786_912_000;
 const PPC_MICROSECONDS_PER_TICK: u64 = 16_625;
 const PPC_QT_GRAPHICS_IMPORTER: u32 = 0x0500_3000;
 const PPC_QT_MOVIE: u32 = 0x0500_3010;
-const PPC_QT_FALLBACK_MOVIE_TASKS_UNTIL_DONE: u32 = 3;
 const PPC_QT_MOVIE_TASKS_PER_SECOND: u64 = 60;
 
 fn ppc_interrupt_callback_stack_pointer(interrupted_sp: u32) -> u32 {
@@ -1855,111 +1856,8 @@ pub enum PpcImportDispatcherTarget {
     Unsupported,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcImportBinding {
-    pub library_index: u32,
-    pub symbol_index: u32,
-    pub library_name: String,
-    pub symbol_name: String,
-    pub class: u8,
-    pub weak: bool,
-    pub address: u32,
-    pub tvector_address: Option<u32>,
-    pub trap_pc: u32,
-    pub dispatcher_target: PpcImportDispatcherTarget,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcStartupProbe {
-    pub result: PpcRunResult,
-    pub first_import_index: Option<u32>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcHleImportTraceEntry {
-    pub import_index: u32,
-    pub library_name: String,
-    pub symbol_name: String,
-    pub pc: u32,
-    pub lr: u32,
-    pub rtoc: u32,
-    pub sp: u32,
-    pub dispatcher_target: PpcImportDispatcherTarget,
-    pub repeat_count: u64,
-}
-
 /// Backward-compatible native-loader name for the shared event record.
 pub type PpcQueuedEvent = QueuedEvent;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcInputSprocketSimpleStateTraceEntry {
-    pub import_index: u32,
-    pub pc: u32,
-    pub element: u32,
-    pub state_ptr: u32,
-    pub state: u32,
-    pub kind: u32,
-    pub kind_name: String,
-    pub fallback_state: u32,
-    pub need_name: String,
-    pub action_binding: String,
-    pub input: PpcInputSnapshot,
-    pub input_sprocket: PpcInputSprocketState,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcDrawSprocketTraceEntry {
-    pub import_index: u32,
-    pub pc: u32,
-    pub action: String,
-    pub result: i16,
-    pub context: Option<u32>,
-    pub requested_state: Option<String>,
-    pub requested_frequency: Option<u32>,
-    pub requested_width: Option<u32>,
-    pub requested_height: Option<u32>,
-    pub requested_context_options: Option<u32>,
-    pub requested_display_depth_mask: Option<u32>,
-    pub requested_back_buffer_depth_mask: Option<u32>,
-    pub requested_display_depth: Option<u32>,
-    pub requested_back_buffer_depth: Option<u32>,
-    pub requested_page_count: Option<u32>,
-    pub can_user_select: Option<bool>,
-    pub fade_kind: Option<String>,
-    pub fade_percent: Option<i32>,
-    pub fade_zero_red: Option<u16>,
-    pub fade_zero_green: Option<u16>,
-    pub fade_zero_blue: Option<u16>,
-    pub reserved_context: Option<u32>,
-    pub active_context: Option<u32>,
-    pub context_state: String,
-    pub front_buffer_gworld: u32,
-    pub back_buffer_gworld: u32,
-    pub last_swap_context: Option<u32>,
-    pub swap_count: u32,
-    pub fade_count: u32,
-    pub frequency: u32,
-    pub width: u32,
-    pub height: u32,
-    pub context_options: u32,
-    pub display_depth_mask: u32,
-    pub back_buffer_depth_mask: u32,
-    pub display_depth: u32,
-    pub back_buffer_depth: u32,
-    pub page_count: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcHleRunProbe {
-    pub result: PpcRunResult,
-    pub handled_import_count: u32,
-    pub last_import_index: Option<u32>,
-    pub unsupported_import_index: Option<u32>,
-    pub import_trace: Vec<PpcHleImportTraceEntry>,
-    pub draw_sprocket_trace: Vec<PpcDrawSprocketTraceEntry>,
-    pub input_sprocket_trace: Vec<PpcInputSprocketSimpleStateTraceEntry>,
-    pub fetch_histogram: Option<PpcFetchHistogram>,
-}
 
 struct PpcHleFetchObserver<'a> {
     histogram: Option<&'a mut PpcFetchHistogram>,
@@ -2851,27 +2749,6 @@ pub(crate) struct PpcAppleEventState {
     pending_dispatches: Vec<PpcAppleEventDispatchAllocation>,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct PpcInputSnapshot {
-    pub key_map: [u8; 16],
-    pub mouse_button: bool,
-    pub mouse_v: i16,
-    pub mouse_h: i16,
-}
-
-impl PpcInputSnapshot {
-    fn key_down(&self, key_code: u8) -> bool {
-        key_map_key_is_down(&self.key_map, key_code)
-    }
-
-    fn any_key_down(&self, key_codes: &[u8]) -> bool {
-        key_codes.iter().copied().any(|key| self.key_down(key))
-    }
-
-    fn is_idle(&self) -> bool {
-        !self.mouse_button && self.key_map.iter().all(|byte| *byte == 0)
-    }
-}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 struct PpcPopUpMenuCall {
@@ -3192,239 +3069,9 @@ fn ppc_restore_menu_definition_port(
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsDirectory {
-    pub dir_id: u32,
-    pub parent_dir_id: u32,
-    pub path: String,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-    pub dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsVolumeRecord {
-    pub ref_num: i16,
-    pub name: String,
-    pub root_dir_id: u32,
-    pub attributes: u16,
-    pub file_count: u16,
-    pub allocation_block_count: u16,
-    pub allocation_block_size: u32,
-    pub clump_size: u32,
-    pub free_blocks: u16,
-    pub bitmap_start: u16,
-    pub allocation_pointer: u16,
-    pub allocation_start: u16,
-    pub next_catalog_id: u32,
-    pub created_date: u32,
-    pub modified_date: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcCfmConnection {
-    pub id: u32,
-    pub library_name: String,
-    pub main_addr: u32,
-    pub init_addr: u32,
-    pub term_addr: u32,
-    pub exports: Vec<PpcCfmExport>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcCfmExport {
-    pub name: String,
-    pub class: u8,
-    pub address: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcCfmLibraryFragment {
-    pub name: String,
-    pub bytes: Vec<u8>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct PpcPreparedMemFragment {
-    main_addr: u32,
-    init_addr: u32,
-    term_addr: u32,
-    exports: Vec<PpcCfmExport>,
-}
-
 pub type PpcHandleRecord = ProcessHandleRecord;
 pub type PpcPtrRecord = ProcessPtrRecord;
 pub type PpcHandleStateRecord = ProcessHandleStateRecord;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcControlRecord {
-    pub handle: u32,
-    pub proc_id: i16,
-    pub popup_menu_id: i16,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcAliasRecord {
-    pub handle: u32,
-    pub target_vref: i16,
-    pub target_dir_id: u32,
-    pub target_name: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcGWorldRecord {
-    pub port: u32,
-    pub pixmap_handle: u32,
-    pub pixmap: u32,
-    pub base_addr: u32,
-    pub gdevice: u32,
-    pub width: u32,
-    pub height: u32,
-    pub depth: u32,
-    pub row_bytes: u32,
-    pub pixels_locked: bool,
-    pub pixels_no_purge: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PpcGWorldAllocationRecord {
-    storage_ptr: u32,
-    pixel_ptr: u32,
-    origin_base: u32,
-    pixel_capacity: u32,
-    ctable_handle: u32,
-    allocation_end: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcFrontBuffer {
-    pub base_addr: u32,
-    pub row_bytes: u32,
-    pub width: u32,
-    pub height: u32,
-    pub depth: u32,
-}
-
-impl From<PpcFrontBuffer> for MenuTrackingSurface {
-    fn from(front: PpcFrontBuffer) -> Self {
-        Self {
-            base_addr: front.base_addr,
-            row_bytes: front.row_bytes,
-            width: front.width,
-            height: front.height,
-            depth: front.depth,
-        }
-    }
-}
-
-impl From<MenuTrackingSurface> for PpcFrontBuffer {
-    fn from(surface: MenuTrackingSurface) -> Self {
-        Self {
-            base_addr: surface.base_addr,
-            row_bytes: surface.row_bytes,
-            width: surface.width,
-            height: surface.height,
-            depth: surface.depth,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PpcQuickDrawSurface {
-    front_buffer: PpcFrontBuffer,
-    top: i16,
-    left: i16,
-    ctable_handle: Option<u32>,
-}
-
-impl PpcQuickDrawSurface {
-    fn local_point(self, (h, v): (i32, i32)) -> (i32, i32) {
-        (h - i32::from(self.left), v - i32::from(self.top))
-    }
-
-    fn local_rect(self, (top, left, bottom, right): (i16, i16, i16, i16)) -> (i32, i32, i32, i32) {
-        (
-            i32::from(top) - i32::from(self.top),
-            i32::from(left) - i32::from(self.left),
-            i32::from(bottom) - i32::from(self.top),
-            i32::from(right) - i32::from(self.left),
-        )
-    }
-
-    fn local_rect_i16(
-        self,
-        (top, left, bottom, right): (i16, i16, i16, i16),
-    ) -> (i16, i16, i16, i16) {
-        (
-            top.wrapping_sub(self.top),
-            left.wrapping_sub(self.left),
-            bottom.wrapping_sub(self.top),
-            right.wrapping_sub(self.left),
-        )
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PpcQ3ViewportRect {
-    left: u32,
-    top: u32,
-    right: u32,
-    bottom: u32,
-}
-
-impl PpcQ3ViewportRect {
-    fn full(front_buffer: PpcFrontBuffer) -> Self {
-        Self {
-            left: 0,
-            top: 0,
-            right: front_buffer.width,
-            bottom: front_buffer.height,
-        }
-    }
-
-    fn from_q3_area(
-        front_buffer: PpcFrontBuffer,
-        min_x: f32,
-        min_y: f32,
-        max_x: f32,
-        max_y: f32,
-    ) -> Option<Self> {
-        if !min_x.is_finite() || !min_y.is_finite() || !max_x.is_finite() || !max_y.is_finite() {
-            return None;
-        }
-        if max_x <= min_x || max_y <= min_y {
-            return None;
-        }
-        let width = front_buffer.width as f32;
-        let height = front_buffer.height as f32;
-        let left = min_x.floor().clamp(0.0, width) as u32;
-        let right = max_x.ceil().clamp(0.0, width) as u32;
-        let top = min_y.floor().clamp(0.0, height) as u32;
-        let bottom = max_y.ceil().clamp(0.0, height) as u32;
-        if right <= left || bottom <= top {
-            return None;
-        }
-        Some(Self {
-            left,
-            top,
-            right,
-            bottom,
-        })
-    }
-
-    fn inclusive_bounds(self) -> Option<(i32, i32, i32, i32)> {
-        if self.right <= self.left || self.bottom <= self.top {
-            return None;
-        }
-        Some((
-            i32::try_from(self.left).ok()?,
-            i32::try_from(self.top).ok()?,
-            i32::try_from(self.right.checked_sub(1)?).ok()?,
-            i32::try_from(self.bottom.checked_sub(1)?).ok()?,
-        ))
-    }
-}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 struct PpcQ3SoftwareFrontBufferSurface {
@@ -3484,48 +3131,6 @@ impl PpcQ3SoftwareFrontBufferSurface {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PpcQ3RenderTargetSource {
-    PixmapDrawContext,
-    MacDrawContext,
-    CurrentGWorld,
-}
-
-impl PpcQ3RenderTargetSource {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::PixmapDrawContext => "pixmap_draw_context",
-            Self::MacDrawContext => "mac_draw_context",
-            Self::CurrentGWorld => "current_gworld",
-        }
-    }
-
-    fn from_str(value: &'static str) -> Option<Self> {
-        match value {
-            "pixmap_draw_context" => Some(Self::PixmapDrawContext),
-            "mac_draw_context" => Some(Self::MacDrawContext),
-            "current_gworld" => Some(Self::CurrentGWorld),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct PpcQ3RenderTarget {
-    front_buffer: PpcFrontBuffer,
-    viewport: Option<PpcQ3ViewportRect>,
-    clear_color: Option<u16>,
-    source: PpcQ3RenderTargetSource,
-    draw_context: Option<u32>,
-    gworld: Option<u32>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3StateOnlyCompletedFrameBatch {
-    frames: usize,
-    target: Option<PpcQ3RenderTarget>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 struct PpcPixMapBits {
     base_addr: u32,
     row_bytes: u32,
@@ -3536,1416 +3141,6 @@ struct PpcPixMapBits {
     width: u32,
     height: u32,
     depth: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcDspContextAttributes {
-    pub frequency: u32,
-    pub width: u32,
-    pub height: u32,
-    pub context_options: u32,
-    pub display_best_depth_mask: u32,
-    pub back_buffer_best_depth_mask: u32,
-    pub display_depth: u32,
-    pub back_buffer_depth: u32,
-    pub page_count: u32,
-}
-
-impl Default for PpcDspContextAttributes {
-    fn default() -> Self {
-        Self {
-            frequency: PPC_DSP_FREQUENCY_60HZ,
-            width: PPC_DSP_SCREEN_WIDTH,
-            height: PPC_DSP_SCREEN_HEIGHT,
-            context_options: PPC_DSP_CONTEXT_OPTION_QD3D_ACCEL,
-            display_best_depth_mask: PPC_DSP_DEPTH_MASK_16,
-            back_buffer_best_depth_mask: PPC_DSP_DEPTH_MASK_16,
-            display_depth: PPC_MAIN_SCREEN_STORAGE_DEPTH,
-            back_buffer_depth: PPC_MAIN_SCREEN_STORAGE_DEPTH,
-            page_count: PPC_DSP_ADVERTISED_PAGE_COUNT,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcRgbColor {
-    pub red: u16,
-    pub green: u16,
-    pub blue: u16,
-}
-
-const PPC_RGB_BLACK: PpcRgbColor = PpcRgbColor {
-    red: 0,
-    green: 0,
-    blue: 0,
-};
-const PPC_RGB_WHITE: PpcRgbColor = PpcRgbColor {
-    red: 0xffff,
-    green: 0xffff,
-    blue: 0xffff,
-};
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PpcDspGammaFadeKind {
-    Manual,
-    In,
-    Out,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PpcDspContextPlayState {
-    Active,
-    Paused,
-    Inactive,
-}
-
-impl Default for PpcDspContextPlayState {
-    fn default() -> Self {
-        Self::Inactive
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcDrawSprocketState {
-    pub started: bool,
-    pub blanking_color: PpcRgbColor,
-    pub reserved_context: Option<u32>,
-    pub active_context: Option<u32>,
-    pub context_state: PpcDspContextPlayState,
-    pub context_attributes: PpcDspContextAttributes,
-    pub front_buffer_gworld: u32,
-    pub back_buffer_gworld: u32,
-    pub last_fade_context: Option<u32>,
-    pub last_fade_kind: Option<PpcDspGammaFadeKind>,
-    pub last_fade_percent: Option<i32>,
-    pub last_fade_zero_color: Option<PpcRgbColor>,
-    pub fade_count: u32,
-    pub last_user_select_display_id: Option<u32>,
-    pub last_user_select_event_proc: Option<u32>,
-    pub user_select_count: u32,
-    pub last_swap_context: Option<u32>,
-    pub swap_count: u32,
-}
-
-impl Default for PpcDrawSprocketState {
-    fn default() -> Self {
-        Self {
-            started: false,
-            blanking_color: PPC_RGB_BLACK,
-            reserved_context: None,
-            active_context: None,
-            context_state: PpcDspContextPlayState::Inactive,
-            context_attributes: PpcDspContextAttributes::default(),
-            front_buffer_gworld: PPC_MAIN_GWORLD,
-            back_buffer_gworld: PPC_DSP_BACK_GWORLD,
-            last_fade_context: None,
-            last_fade_kind: None,
-            last_fade_percent: None,
-            last_fade_zero_color: None,
-            fade_count: 0,
-            last_user_select_display_id: None,
-            last_user_select_event_proc: None,
-            user_select_count: 0,
-            last_swap_context: None,
-            swap_count: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PpcQ3ObjectKind {
-    Generic,
-    MemoryStorage,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct PpcQ3ObjectSource {
-    pub file: u32,
-    pub offset: u32,
-    pub parent_group_type: u32,
-    pub group_depth: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3ObjectRecord {
-    pub object: u32,
-    pub kind: PpcQ3ObjectKind,
-    pub object_type: u32,
-    pub source: PpcQ3ObjectSource,
-    pub data_ptr: u32,
-    pub data_size: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3ObjectReferenceRecord {
-    pub object: u32,
-    pub ref_count: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3ErrorState {
-    pub first_error: u32,
-    pub last_error: u32,
-    pub clear_on_next_q3_call: bool,
-}
-
-impl Default for PpcQ3ErrorState {
-    fn default() -> Self {
-        Self {
-            first_error: PPC_Q3_ERROR_NONE,
-            last_error: PPC_Q3_ERROR_NONE,
-            clear_on_next_q3_call: false,
-        }
-    }
-}
-
-impl PpcQ3ErrorState {
-    fn clear(&mut self) {
-        self.first_error = PPC_Q3_ERROR_NONE;
-        self.last_error = PPC_Q3_ERROR_NONE;
-        self.clear_on_next_q3_call = false;
-    }
-
-    fn post(&mut self, error: u32) {
-        if error == PPC_Q3_ERROR_NONE {
-            return;
-        }
-        if self.first_error == PPC_Q3_ERROR_NONE {
-            self.first_error = error;
-        }
-        self.last_error = error;
-        self.clear_on_next_q3_call = false;
-    }
-
-    fn get(&mut self) -> (u32, u32) {
-        let errors = (self.first_error, self.last_error);
-        if self.last_error != PPC_Q3_ERROR_NONE {
-            self.clear_on_next_q3_call = true;
-        }
-        errors
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct PpcQ3LifecycleState {
-    pub initialize_count: u32,
-    pub exit_count: u32,
-    pub initialized_depth: u32,
-}
-
-impl PpcQ3LifecycleState {
-    pub fn initialized(&self) -> bool {
-        self.initialized_depth != 0
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3MemoryStorageRecord {
-    pub storage: u32,
-    pub buffer_ptr: u32,
-    pub valid_size: u32,
-    pub buffer_size: u32,
-    pub owns_buffer: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3FileRecord {
-    pub file: u32,
-    pub storage: u32,
-    pub is_open: bool,
-    pub object_type: u32,
-    pub read_offset: u32,
-    pub read_object: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3GroupMembershipRecord {
-    pub group: u32,
-    pub object: u32,
-    pub before: Option<u32>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3FileGroupRecord {
-    pub file: u32,
-    pub offset: u32,
-    pub group: u32,
-    pub group_type: u32,
-    pub parent_group: u32,
-    pub group_depth: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum PpcQ3SubmissionKind {
-    Shader,
-    Style,
-    FogStyle,
-    TriMesh,
-    MatrixTransform,
-    ResetTransform,
-    Push,
-    Pop,
-    Object,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3SubmissionRecord {
-    pub view: u32,
-    pub kind: PpcQ3SubmissionKind,
-    pub primary: u32,
-    pub secondary: u32,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3ViewTransformRecord {
-    pub view: u32,
-    pub stack: Vec<[[f32; 4]; 4]>,
-    pub local_to_world: [[f32; 4]; 4],
-}
-
-impl PpcQ3ViewTransformRecord {
-    fn new(view: u32) -> Self {
-        Self {
-            view,
-            stack: Vec::new(),
-            local_to_world: ppc_q3_matrix4x4_identity(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SubmissionTransformRecord {
-    pub view: u32,
-    pub kind: PpcQ3SubmissionKind,
-    pub primary: u32,
-    pub secondary: u32,
-    pub local_to_world: [[f32; 4]; 4],
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3ViewMaterialRecord {
-    pub view: u32,
-    pub shader: u32,
-    pub illumination_type: u32,
-    pub styles: Vec<PpcQ3StyleRecord>,
-    pub fog_style: Option<PpcQ3FogStyleData>,
-    pub attributes: Vec<PpcQ3AttributeRecord>,
-}
-
-impl PpcQ3ViewMaterialRecord {
-    fn new(view: u32) -> Self {
-        Self {
-            view,
-            shader: 0,
-            illumination_type: PPC_Q3_ILLUMINATION_TYPE_PHONG,
-            styles: Vec::new(),
-            fog_style: None,
-            attributes: Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SubmissionMaterialRecord {
-    pub view: u32,
-    pub kind: PpcQ3SubmissionKind,
-    pub primary: u32,
-    pub secondary: u32,
-    pub shader: u32,
-    pub illumination_type: u32,
-    pub styles: Vec<PpcQ3StyleRecord>,
-    pub fog_style: Option<PpcQ3FogStyleData>,
-    pub attributes: Vec<PpcQ3AttributeRecord>,
-    pub shader_uv_transform: Option<PpcQ3ShaderUvTransformRecord>,
-    pub shader_boundary: Option<PpcQ3ShaderBoundaryRecord>,
-    pub texture_shader: Option<PpcQ3TextureShaderRecord>,
-    pub mipmap_texture: Option<PpcQ3MipmapTextureRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SubmissionLightRecord {
-    pub view: u32,
-    pub kind: PpcQ3SubmissionKind,
-    pub primary: u32,
-    pub secondary: u32,
-    pub light_group: u32,
-    pub lights: Vec<PpcQ3LightRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3ViewStateSnapshotRecord {
-    pub view: u32,
-    pub transform: Option<PpcQ3ViewTransformRecord>,
-    pub material: Option<PpcQ3ViewMaterialRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3CompletedFrameRecord {
-    pub view: u32,
-    pub submissions: Vec<PpcQ3SubmissionRecord>,
-    pub submission_transforms: Vec<PpcQ3SubmissionTransformRecord>,
-    pub submission_materials: Vec<PpcQ3SubmissionMaterialRecord>,
-    pub submission_lights: Vec<PpcQ3SubmissionLightRecord>,
-    pub retained_trimeshes: Vec<PpcQ3TriMeshRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3RetainedFrameRecord {
-    pub view: u32,
-    pub frame: PpcQ3CompletedFrameRecord,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum PpcQ3SceneCommand {
-    TriMesh(PpcQ3SceneTriMeshCommand),
-    Submission(PpcQ3SceneSubmissionCommand),
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SceneTriMeshCommand {
-    pub submission_index: usize,
-    pub view_state: PpcQ3ViewStateRecord,
-    pub camera: Option<PpcQ3CameraRecord>,
-    pub geometry: PpcQ3SceneTriMeshGeometry,
-    pub local_to_world: [[f32; 4]; 4],
-    pub material: PpcQ3SubmissionMaterialRecord,
-    pub lights: PpcQ3SubmissionLightRecord,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SceneSubmissionCommand {
-    pub submission_index: usize,
-    pub view_state: PpcQ3ViewStateRecord,
-    pub submission: PpcQ3SubmissionRecord,
-    pub local_to_world: [[f32; 4]; 4],
-    pub material: PpcQ3SubmissionMaterialRecord,
-    pub lights: PpcQ3SubmissionLightRecord,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3SceneTriMeshGeometry {
-    pub source: PpcQ3SceneTriMeshSource,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum PpcQ3SceneTriMeshSource {
-    Object(u32),
-    DataPtr(u32),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3SceneReplayMemoryRegion {
-    pub base_addr: u32,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3SceneReplay {
-    pub commands: Vec<PpcQ3SceneCommand>,
-    pub memory_regions: Vec<PpcQ3SceneReplayMemoryRegion>,
-}
-
-/// A browser-friendly QD3D frame whose guest geometry has already been
-/// transformed, clipped, lit, and decoded for direct GPU rasterization.
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3GpuFrame {
-    pub width: u32,
-    pub height: u32,
-    pub viewport: [i32; 4],
-    pub clear_color: Option<[f32; 4]>,
-    pub textures: Vec<PpcQ3GpuTexture>,
-    pub draws: Vec<PpcQ3GpuDraw>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQ3GpuTexture {
-    pub width: u32,
-    pub height: u32,
-    pub rgba: Vec<u8>,
-    pub wrap_u: bool,
-    pub wrap_v: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct PpcQ3GpuDraw {
-    pub texture: Option<usize>,
-    pub vertices: Vec<PpcQ3GpuVertex>,
-    pub blend: bool,
-    pub write_depth: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct PpcQ3GpuVertex {
-    pub screen_x: f32,
-    pub screen_y: f32,
-    pub depth: f32,
-    pub reciprocal_w: f32,
-    pub uv: [f32; 2],
-    pub color: [f32; 4],
-}
-
-impl PpcQ3SceneReplay {
-    pub fn to_json_pretty(&self) -> serde_json::Result<String> {
-        serde_json::to_string_pretty(self)
-    }
-
-    pub fn from_json_str(value: &str) -> serde_json::Result<Self> {
-        serde_json::from_str(value)
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct PpcQ3SoftwareRenderStats {
-    pub frames: usize,
-    pub commands: usize,
-    pub vertices: usize,
-    pub triangles: usize,
-    pub pixels: usize,
-    pub target_base: Option<u32>,
-    pub target_row_bytes: Option<u32>,
-    pub target_width: Option<u32>,
-    pub target_height: Option<u32>,
-    pub target_depth: Option<u32>,
-    pub target_source: Option<&'static str>,
-    pub target_draw_context: Option<u32>,
-    pub target_gworld: Option<u32>,
-    pub target_consistent: bool,
-}
-
-impl PpcQ3SoftwareRenderStats {
-    fn add_empty_frames(&mut self, frames: usize) {
-        self.frames = self.frames.saturating_add(frames);
-    }
-
-    fn add_state_only_batches(&mut self, batches: &[PpcQ3StateOnlyCompletedFrameBatch]) {
-        for batch in batches {
-            if let Some(target) = batch.target {
-                self.record_target(target);
-            }
-            self.add_empty_frames(batch.frames);
-        }
-    }
-
-    fn record_target(&mut self, target_record: PpcQ3RenderTarget) {
-        let front_buffer = target_record.front_buffer;
-        let target = (
-            front_buffer.base_addr,
-            front_buffer.row_bytes,
-            front_buffer.width,
-            front_buffer.height,
-            front_buffer.depth,
-            Some(target_record.source.as_str()),
-            target_record.draw_context,
-            target_record.gworld,
-        );
-        let current = (
-            self.target_base,
-            self.target_row_bytes,
-            self.target_width,
-            self.target_height,
-            self.target_depth,
-            self.target_source,
-            self.target_draw_context,
-            self.target_gworld,
-        );
-        if self.target_base.is_none() {
-            self.target_base = Some(target.0);
-            self.target_row_bytes = Some(target.1);
-            self.target_width = Some(target.2);
-            self.target_height = Some(target.3);
-            self.target_depth = Some(target.4);
-            self.target_source = target.5;
-            self.target_draw_context = target.6;
-            self.target_gworld = target.7;
-            self.target_consistent = true;
-        } else if current
-            != (
-                Some(target.0),
-                Some(target.1),
-                Some(target.2),
-                Some(target.3),
-                Some(target.4),
-                target.5,
-                target.6,
-                target.7,
-            )
-        {
-            self.target_consistent = false;
-        }
-    }
-
-    fn merge_frame_stats(
-        &mut self,
-        frame_stats: PpcQ3SoftwareRenderStats,
-        saw_missing_target: &mut bool,
-    ) {
-        self.frames = self.frames.saturating_add(1);
-        self.commands = self.commands.saturating_add(frame_stats.commands);
-        self.vertices = self.vertices.saturating_add(frame_stats.vertices);
-        self.triangles = self.triangles.saturating_add(frame_stats.triangles);
-        self.pixels = self.pixels.saturating_add(frame_stats.pixels);
-        if let (
-            Some(target_base),
-            Some(target_row_bytes),
-            Some(target_width),
-            Some(target_height),
-            Some(target_depth),
-            Some(target_source),
-        ) = (
-            frame_stats.target_base,
-            frame_stats.target_row_bytes,
-            frame_stats.target_width,
-            frame_stats.target_height,
-            frame_stats.target_depth,
-            frame_stats.target_source,
-        ) {
-            if let Some(source) = PpcQ3RenderTargetSource::from_str(target_source) {
-                self.record_target(PpcQ3RenderTarget {
-                    front_buffer: PpcFrontBuffer {
-                        base_addr: target_base,
-                        row_bytes: target_row_bytes,
-                        width: target_width,
-                        height: target_height,
-                        depth: target_depth,
-                    },
-                    viewport: None,
-                    clear_color: None,
-                    source,
-                    draw_context: frame_stats.target_draw_context,
-                    gworld: frame_stats.target_gworld,
-                });
-                self.target_consistent &= frame_stats.target_consistent;
-                if *saw_missing_target {
-                    self.target_consistent = false;
-                }
-            } else {
-                *saw_missing_target = true;
-                if self.target_base.is_some() {
-                    self.target_consistent = false;
-                }
-            }
-        } else {
-            *saw_missing_target = true;
-            if self.target_base.is_some() {
-                self.target_consistent = false;
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3FogStyleData {
-    pub state: u32,
-    pub mode: u32,
-    pub fog_start: f32,
-    pub fog_end: f32,
-    pub density: f32,
-    pub color: (f32, f32, f32, f32),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct PpcQ3FogStyleRecord {
-    pub view: u32,
-    pub data_ptr: u32,
-    pub data: PpcQ3FogStyleData,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3AttributeRecord {
-    pub attribute_set: u32,
-    pub attribute_type: u32,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3ShaderUvTransformRecord {
-    pub shader: u32,
-    pub matrix: [[f32; 3]; 3],
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3ShaderBoundaryRecord {
-    pub shader: u32,
-    pub u_boundary: u32,
-    pub v_boundary: u32,
-}
-
-impl PpcQ3ShaderBoundaryRecord {
-    fn new(shader: u32) -> Self {
-        Self {
-            shader,
-            u_boundary: PPC_Q3_SHADER_UV_BOUNDARY_WRAP,
-            v_boundary: PPC_Q3_SHADER_UV_BOUNDARY_WRAP,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3MipmapTextureRecord {
-    pub texture: u32,
-    pub mipmap: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3TextureShaderRecord {
-    pub shader: u32,
-    pub texture: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQ3RendererPreferenceRecord {
-    pub renderer: u32,
-    pub double_buffer_bypass: Option<u32>,
-    pub preference_vendor: Option<u32>,
-    pub preference_engine: Option<u32>,
-    pub rave_context_hints: Option<u32>,
-    pub rave_texture_filter: Option<u32>,
-}
-
-impl PpcQ3RendererPreferenceRecord {
-    fn new(renderer: u32) -> Self {
-        Self {
-            renderer,
-            double_buffer_bypass: None,
-            preference_vendor: None,
-            preference_engine: None,
-            rave_context_hints: None,
-            rave_texture_filter: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQ3DrawContextRecord {
-    pub draw_context: u32,
-    pub draw_context_type: u32,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQ3TriMeshRecord {
-    pub trimesh: u32,
-    pub data: Vec<u8>,
-    pub triangle_attribute_sets: Vec<u32>,
-    pub get_data_copies: Vec<PpcQ3TriMeshGetDataCopyRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQ3TriMeshGetDataCopyRecord {
-    pub data_out_ptr: u32,
-    pub data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum PpcQ3StyleKind {
-    Backfacing,
-    Interpolation,
-    Fill,
-    Orientation,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3StyleRecord {
-    pub style: u32,
-    pub kind: PpcQ3StyleKind,
-    pub value: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3CameraPlacement {
-    pub camera_location: (f32, f32, f32),
-    pub point_of_interest: (f32, f32, f32),
-    pub up_vector: (f32, f32, f32),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub enum PpcQ3CameraProjection {
-    ViewAngleAspect {
-        fov: f32,
-        aspect_ratio_x_to_y: f32,
-    },
-    Orthographic {
-        left: f32,
-        top: f32,
-        right: f32,
-        bottom: f32,
-    },
-    ViewPlane {
-        view_plane: f32,
-        half_width_at_view_plane: f32,
-        half_height_at_view_plane: f32,
-        center_x_on_view_plane: f32,
-        center_y_on_view_plane: f32,
-    },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-struct PpcQ3CameraCommonData {
-    placement: PpcQ3CameraPlacement,
-    range_hither: f32,
-    range_yon: f32,
-    viewport_origin: (f32, f32),
-    viewport_width: f32,
-    viewport_height: f32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3CameraRecord {
-    pub camera: u32,
-    pub camera_type: u32,
-    pub placement: PpcQ3CameraPlacement,
-    pub range_hither: f32,
-    pub range_yon: f32,
-    pub viewport_origin: (f32, f32),
-    pub viewport_width: f32,
-    pub viewport_height: f32,
-    pub projection: PpcQ3CameraProjection,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3LightData {
-    pub is_on: u32,
-    pub brightness: f32,
-    pub color: (f32, f32, f32),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub enum PpcQ3LightKind {
-    Ambient,
-    Directional {
-        casts_shadows: u32,
-        direction: (f32, f32, f32),
-    },
-    Point {
-        casts_shadows: u32,
-        attenuation: u32,
-        location: (f32, f32, f32),
-    },
-    Spot {
-        casts_shadows: u32,
-        attenuation: u32,
-        location: (f32, f32, f32),
-        direction: (f32, f32, f32),
-        hot_angle: f32,
-        outer_angle: f32,
-        fall_off: u32,
-    },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub struct PpcQ3LightRecord {
-    pub light: u32,
-    pub light_type: u32,
-    pub data: PpcQ3LightData,
-    pub kind: PpcQ3LightKind,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PpcQ3ViewStateRecord {
-    pub view: u32,
-    pub renderer: u32,
-    pub light_group: u32,
-    pub draw_context: u32,
-    pub camera: u32,
-    pub rendering_depth: u32,
-    pub bounding_box_depth: u32,
-    pub cancelled: bool,
-}
-
-impl PpcQ3ViewStateRecord {
-    fn new(view: u32) -> Self {
-        Self {
-            view,
-            renderer: 0,
-            light_group: 0,
-            draw_context: 0,
-            camera: 0,
-            rendering_depth: 0,
-            bounding_box_depth: 0,
-            cancelled: false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcInputSprocketState {
-    pub initialized: bool,
-    pub suspended: bool,
-    pub keyboard_active: bool,
-    pub mouse_active: bool,
-    pub configure_count: u32,
-    pub virtual_element_count: u32,
-    pub last_virtual_need_count: u32,
-    pub last_virtual_needs_ptr: u32,
-    pub last_virtual_elements_out_ptr: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PpcInputSprocketActionBinding {
-    AxisYaw,
-    AxisPitch,
-    AxisHorizontal,
-    AxisVertical,
-    AxisDirectional,
-    ButtonLeft,
-    ButtonRight,
-    ButtonForward,
-    ButtonBackward,
-    ButtonCameraLeft,
-    ButtonCameraRight,
-    ButtonJump,
-    ButtonFire,
-    ButtonWeapon,
-    ButtonPickup,
-    ButtonJetUp,
-    ButtonJetDown,
-    ButtonPause,
-    ButtonZoomIn,
-    ButtonZoomOut,
-    ButtonCameraMode,
-    ButtonToggleMusic,
-    ButtonToggleAmbientSound,
-    ButtonVolumeUp,
-    ButtonVolumeDown,
-    ButtonToggleGps,
-    ButtonQuit,
-    ButtonPrimary,
-    DpadDirectional,
-    DeltaYaw,
-    DeltaPitch,
-    Unknown,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcInputSprocketVirtualElementRecord {
-    pub element: u32,
-    pub need_index: u32,
-    pub need_source: u32,
-    pub kind: u32,
-    pub default_state: u32,
-    pub action_binding: PpcInputSprocketActionBinding,
-    pub need_name: String,
-    pub need_record: Vec<u8>,
-}
-
-struct PpcInputSprocketVirtualElementDraft {
-    need_index: u32,
-    need_source: u32,
-    kind: u32,
-    default_state: u32,
-    action_binding: PpcInputSprocketActionBinding,
-    need_name: String,
-    need_record: Vec<u8>,
-}
-
-impl Default for PpcInputSprocketState {
-    fn default() -> Self {
-        Self {
-            initialized: false,
-            suspended: false,
-            keyboard_active: true,
-            mouse_active: true,
-            configure_count: 0,
-            virtual_element_count: 0,
-            last_virtual_need_count: 0,
-            last_virtual_needs_ptr: 0,
-            last_virtual_elements_out_ptr: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQuickTimeVideoTrackRecord {
-    pub media_time_scale: u32,
-    pub media_duration: u64,
-    pub sample_count: u32,
-    pub first_sample_duration: u32,
-    pub first_sample_size: u32,
-    pub first_chunk_offset: u64,
-    pub first_sample_data_len: u32,
-    pub first_sample_checksum: u32,
-    pub first_sample_preview_len: u8,
-    pub first_sample_preview: [u8; 16],
-    pub first_samples_per_chunk: u32,
-    pub sample_description_id: u32,
-    pub codec: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQuickTimeVideoSampleRecord {
-    pub offset: u64,
-    pub size: u32,
-    pub media_start_time: u64,
-    pub duration: u32,
-    pub data_len: u32,
-    pub checksum: u32,
-    pub preview_len: u8,
-    pub preview: [u8; 16],
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQuickTimeVideoSampleTableRecord {
-    pub media_time_scale: u32,
-    pub media_duration: u64,
-    pub sample_count: u32,
-    pub codec: u32,
-    pub samples: Vec<PpcQuickTimeVideoSampleRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQuickTimeVideoDecodeCacheRecord {
-    codec: u32,
-    width: usize,
-    height: usize,
-    sample_index: usize,
-    rgb: Vec<u8>,
-    cinepak_strips: Vec<PpcQuickTimeCinepakStripState>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcQuickTimeAudioTrackRecord {
-    pub media_time_scale: u32,
-    pub media_duration: u64,
-    pub sample_count: u32,
-    pub first_sample_duration: u32,
-    pub first_sample_size: u32,
-    pub first_chunk_offset: u64,
-    pub first_sample_data_len: u32,
-    pub first_sample_checksum: u32,
-    pub first_sample_preview_len: u8,
-    pub first_sample_preview: [u8; 16],
-    pub first_samples_per_chunk: u32,
-    pub sample_description_id: u32,
-    pub codec: u32,
-    pub channel_count: u16,
-    pub sample_size_bits: u16,
-    pub sample_rate_fixed: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcQuickTimeState {
-    pub movie_toolbox_enter_count: u32,
-    pub movie_toolbox_exit_count: u32,
-    pub movie_toolbox_init_depth: u32,
-    pub movie_error: i16,
-    pub movie_sticky_error: i16,
-    pub movie_file_open_count: u32,
-    pub movie_file_close_count: u32,
-    pub movie_file_ref_num: i16,
-    pub movie_file_last_closed_ref_num: i16,
-    pub movie_file_path: String,
-    pub movie_file_data: Vec<u8>,
-    pub movie_file_bounds: Option<(i16, i16, i16, i16)>,
-    pub movie_file_time_scale: u32,
-    pub movie_file_duration: u64,
-    pub movie_file_tasks_until_done: u32,
-    pub movie_file_video_track: Option<PpcQuickTimeVideoTrackRecord>,
-    pub movie_file_video_samples: Option<PpcQuickTimeVideoSampleTableRecord>,
-    pub movie_file_audio_track: Option<PpcQuickTimeAudioTrackRecord>,
-    pub graphics_importer_gworld: u32,
-    pub graphics_importer_gdevice: u32,
-    pub graphics_importer_open: bool,
-    pub graphics_import_draw_count: u32,
-    pub graphics_import_source_draw_count: u32,
-    pub graphics_importer_path: String,
-    pub graphics_importer_data: Vec<u8>,
-    pub graphics_importer_bounds: Option<(i16, i16, i16, i16)>,
-    pub movie_gworld: u32,
-    pub movie_gdevice: u32,
-    pub movie_box: (i16, i16, i16, i16),
-    pub movie_set_box_count: u32,
-    pub movie_beginning_count: u32,
-    pub movie_at_beginning: bool,
-    pub movie_started: bool,
-    pub movie_task_count: u32,
-    pub movie_tasks_until_done: u32,
-    pub movie_video_track: Option<PpcQuickTimeVideoTrackRecord>,
-    pub movie_video_samples: Option<PpcQuickTimeVideoSampleTableRecord>,
-    pub movie_video_decode_cache: Option<PpcQuickTimeVideoDecodeCacheRecord>,
-    pub movie_audio_track: Option<PpcQuickTimeAudioTrackRecord>,
-    pub movie_disposed: bool,
-    pub movie_volume: i16,
-    pub movie_time_base_flags: u32,
-}
-
-impl Default for PpcQuickTimeState {
-    fn default() -> Self {
-        Self {
-            movie_toolbox_enter_count: 0,
-            movie_toolbox_exit_count: 0,
-            movie_toolbox_init_depth: 0,
-            movie_error: PPC_NO_ERR,
-            movie_sticky_error: PPC_NO_ERR,
-            movie_file_open_count: 0,
-            movie_file_close_count: 0,
-            movie_file_ref_num: 0,
-            movie_file_last_closed_ref_num: 0,
-            movie_file_path: String::new(),
-            movie_file_data: Vec::new(),
-            movie_file_bounds: None,
-            movie_file_time_scale: 0,
-            movie_file_duration: 0,
-            movie_file_tasks_until_done: PPC_QT_FALLBACK_MOVIE_TASKS_UNTIL_DONE,
-            movie_file_video_track: None,
-            movie_file_video_samples: None,
-            movie_file_audio_track: None,
-            graphics_importer_gworld: 0,
-            graphics_importer_gdevice: 0,
-            graphics_importer_open: false,
-            graphics_import_draw_count: 0,
-            graphics_import_source_draw_count: 0,
-            graphics_importer_path: String::new(),
-            graphics_importer_data: Vec::new(),
-            graphics_importer_bounds: None,
-            movie_gworld: 0,
-            movie_gdevice: 0,
-            movie_box: (
-                0,
-                0,
-                PPC_MAIN_SCREEN_HEIGHT as i16,
-                PPC_MAIN_SCREEN_WIDTH as i16,
-            ),
-            movie_set_box_count: 0,
-            movie_beginning_count: 0,
-            movie_at_beginning: true,
-            movie_started: false,
-            movie_task_count: 0,
-            movie_tasks_until_done: PPC_QT_FALLBACK_MOVIE_TASKS_UNTIL_DONE,
-            movie_video_track: None,
-            movie_video_samples: None,
-            movie_video_decode_cache: None,
-            movie_audio_track: None,
-            movie_disposed: false,
-            movie_volume: 0x0100,
-            movie_time_base_flags: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSndCommandRecord {
-    pub channel: u32,
-    pub command: u16,
-    pub param1: i16,
-    pub param2: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSoundFilePlaybackRecord {
-    pub channel: u32,
-    pub ref_num: i16,
-    pub resource_id: i16,
-    pub buffer_size: u32,
-    pub buffer: u32,
-    pub selection: u32,
-    pub completion: u32,
-    pub completion_command: Option<PpcSndCommandRecord>,
-    pub async_play: bool,
-    pub paused: bool,
-    pub active: bool,
-    pub quiet_now: bool,
-    pub timing_valid: bool,
-    pub start_tick: u32,
-    pub start_instruction_count: u64,
-    pub due_tick: u32,
-    pub due_instruction_count: u64,
-    pub pause_timing_valid: bool,
-    pub pause_started_tick: u32,
-    pub pause_started_instruction_count: u64,
-    pub aiff: Option<PpcAiffMetadata>,
-    pub decoded_aiff: Option<PpcDecodedAiffSamples>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct PpcSoundState {
-    pub queued_commands: Vec<PpcSndCommandRecord>,
-    pub immediate_commands: Vec<PpcSndCommandRecord>,
-    pub decoded_buffer_commands: Vec<PpcDecodedBufferCommandRecord>,
-    pub double_buffer_playbacks: Vec<PpcSoundDoubleBufferPlaybackRecord>,
-    pub file_playbacks: Vec<PpcSoundFilePlaybackRecord>,
-    pub decoded_file_playbacks: Vec<PpcDecodedAiffPlaybackRecord>,
-    pub pending_completions: Vec<PpcSoundCompletionRecord>,
-    pub pending_doublebacks: Vec<PpcSoundDoubleBackRecord>,
-    pub completion_invocations: Vec<PpcSoundCompletionInvocationRecord>,
-    pub sys_beep_count: u32,
-    pub last_sys_beep_duration: i16,
-    pub start_count: u32,
-    pub pause_count: u32,
-    pub stop_count: u32,
-    pub double_buffer_play_count: u32,
-    pub last_double_buffer_channel: u32,
-    pub last_double_buffer_header: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcDecodedBufferCommandRecord {
-    pub channel: u32,
-    pub sample_rate_fixed: u32,
-    pub samples: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSoundDoubleBufferPlaybackRecord {
-    pub channel: u32,
-    pub header: u32,
-    pub buffers: [u32; 2],
-    pub callback: u32,
-    pub sample_rate_fixed: u32,
-    pub num_channels: u16,
-    pub sample_size: u16,
-    pub compression_id: i16,
-    pub packet_size: u16,
-    pub current_buffer_index: u8,
-    pub callback_pending_mask: u8,
-    pub active: bool,
-    pub host_initialized: bool,
-    pub host_buffer_loaded: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSoundDoubleBackRecord {
-    pub channel: u32,
-    pub header: u32,
-    pub exhausted_buffer: u32,
-    pub exhausted_buffer_index: u32,
-    pub callback: u32,
-    pub tick: u32,
-    pub instruction_count: u64,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcVblTaskRecord {
-    pub task_ptr: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcTimerTaskRecord {
-    pub task_ptr: u32,
-    pub callback: u32,
-    pub active: bool,
-    pub fire_at_tick: u32,
-    pub last_fired_tick: Option<u32>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcTimerCallbackInvocationRecord {
-    pub task_ptr: u32,
-    pub callback: u32,
-    pub callback_entry: u32,
-    pub callback_rtoc: u32,
-    pub tick: u32,
-    pub cycles: u64,
-    pub end_pc: u32,
-    pub end_sp: u32,
-    pub end_r3: u32,
-    pub result: PpcRunResult,
-    pub unsupported_import_index: Option<u32>,
-}
-
-#[derive(Debug, Clone)]
-pub struct PpcTimerCallbackProbe {
-    pub invocation: PpcTimerCallbackInvocationRecord,
-    pub import_trace: Vec<PpcHleImportTraceEntry>,
-    pub fetch_histogram: Option<PpcFetchHistogram>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcVblCallbackInvocationRecord {
-    pub task_ptr: u32,
-    pub callback: u32,
-    pub callback_entry: u32,
-    pub callback_rtoc: u32,
-    pub tick: u32,
-    pub cycles: u64,
-    pub end_pc: u32,
-    pub end_sp: u32,
-    pub end_r3: u32,
-    pub result: PpcRunResult,
-    pub unsupported_import_index: Option<u32>,
-}
-
-#[derive(Debug, Clone)]
-pub struct PpcVblCallbackProbe {
-    pub invocation: PpcVblCallbackInvocationRecord,
-    pub import_trace: Vec<PpcHleImportTraceEntry>,
-    pub fetch_histogram: Option<PpcFetchHistogram>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcAiffMetadata {
-    pub form_type: u32,
-    pub channel_count: u16,
-    pub sample_frame_count: u32,
-    pub sample_size: u16,
-    pub sample_rate_hz: u32,
-    pub compression_type: u32,
-    pub sound_data_offset: u32,
-    pub sound_data_size: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcDecodedAiffSamples {
-    pub sample_rate_fixed: u32,
-    pub sample_count: u32,
-    pub preview_len: u8,
-    pub preview: [u8; 16],
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcDecodedAiffPlaybackRecord {
-    pub file_playback_index: u32,
-    pub channel: u32,
-    pub sample_rate_fixed: u32,
-    pub samples: Vec<u8>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSoundCompletionRecord {
-    pub file_playback_index: u32,
-    pub channel: u32,
-    pub completion: u32,
-    pub command: Option<PpcSndCommandRecord>,
-    pub tick: u32,
-    pub instruction_count: u64,
-    pub scheduled_tick: u32,
-    pub scheduled_instruction_count: u64,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PpcSoundCompletionInvocationRecord {
-    pub file_playback_index: u32,
-    pub channel: u32,
-    pub completion: u32,
-    pub callback_entry: u32,
-    pub callback_rtoc: u32,
-    pub tick: u32,
-    pub instruction_count: u64,
-    pub scheduled_tick: u32,
-    pub scheduled_instruction_count: u64,
-    pub cycles: u64,
-    pub end_pc: u32,
-    pub end_sp: u32,
-    pub end_r3: u32,
-    pub result: PpcRunResult,
-    pub unsupported_import_index: Option<u32>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcSoundCompletionCallProbe {
-    pub invocation: PpcSoundCompletionInvocationRecord,
-    pub import_trace: Vec<PpcHleImportTraceEntry>,
-    pub fetch_histogram: Option<PpcFetchHistogram>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcFileRecord {
-    pub ref_num: i16,
-    pub path: String,
-    pub position: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct PpcStdioStreamRecord {
-    ref_num: Option<i16>,
-    path: Option<String>,
-    position: u32,
-    standard: bool,
-    readable: bool,
-    writable: bool,
-    append: bool,
-    closed: bool,
-    eof: bool,
-    error: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsFileRecord {
-    pub path: String,
-    pub data: Vec<u8>,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-    pub dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcResourceFileRecord {
-    pub ref_num: i16,
-    pub path: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsResourceFileRecord {
-    pub path: String,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-    pub resource_len: u32,
-    pub raw_data: Option<Vec<u8>>,
-    pub map_attrs: u16,
-    pub dirty: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsResourceForkExport {
-    pub path: String,
-    pub data: Vec<u8>,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsFileExport {
-    pub path: String,
-    pub data: Vec<u8>,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsDirectoryExport {
-    pub path: String,
-    pub creator: u32,
-    pub file_type: u32,
-    pub finder_flags: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PpcVfsResourceRecord {
-    pub ref_num: i16,
-    pub path: String,
-    pub res_type: u32,
-    pub res_id: i16,
-    pub name: Vec<u8>,
-    pub data: Vec<u8>,
-    pub raw_data: Option<Vec<u8>>,
-    pub raw_attrs: Option<u16>,
-    pub attrs: u16,
-    pub handle: u32,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct PpcScrapState {
-    private_text_handle: u32,
-    desktop_initialized: bool,
-    desktop_flavors: Vec<(u32, Vec<u8>)>,
-}
-
-#[derive(Debug, Clone)]
-struct PpcListRecord {
-    handle: u32,
-    cells_handle: u32,
-    data_bounds: (i16, i16, i16, i16),
-    cells: Vec<Vec<u8>>,
-    selected: Vec<bool>,
-    draw_enabled: bool,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct PpcListManagerState {
-    lists: Vec<PpcListRecord>,
 }
 
 #[derive(Debug, Default)]
@@ -43705,13 +41900,7 @@ fn ppc_write_q3_matrix4x4(
     Some(())
 }
 
-fn ppc_q3_matrix4x4_identity() -> [[f32; 4]; 4] {
-    let mut matrix = [[0.0; 4]; 4];
-    for (index, row) in matrix.iter_mut().enumerate() {
-        row[index] = 1.0;
-    }
-    matrix
-}
+
 
 fn ppc_q3_matrix3x3_identity() -> [[f32; 3]; 3] {
     let mut matrix = [[0.0; 3]; 3];
@@ -49822,20 +48011,7 @@ fn ppc_qt_decode_ima4_channel_packet(packet: &[u8]) -> Option<[i16; 64]> {
     Some(decoded)
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct PpcQuickTimeCinepakStripState {
-    v4_codebook: [[u8; 12]; 256],
-    v1_codebook: [[u8; 12]; 256],
-}
 
-impl Default for PpcQuickTimeCinepakStripState {
-    fn default() -> Self {
-        Self {
-            v4_codebook: [[0; 12]; 256],
-            v1_codebook: [[0; 12]; 256],
-        }
-    }
-}
 
 struct PpcQuickTimeCinepakDecoder {
     width: usize,
